Subida CRUD generos

+++ app/modelos/generos.modelo.php
@@ -6,7 +6,7 @@ class GenerosModelo {
     public function __construct() {
        $this->db = new PDO('mysql:host=localhost;dbname=db_tpe;charset=utf8', 'root', '');
     }
- 
+    //Revisado
     public function obtenerGeneros() {
         // 2. Ejecuto la consulta
         $query = $this->db->prepare('SELECT * FROM generos');
@@ -14,7 +14,6 @@ class GenerosModelo {
     
         // 3. Obtengo los datos en un arreglo de objetos
         $generos = $query->fetchAll(PDO::FETCH_OBJ); 
-        var_dump($generos);
         return $generos;
       }
  
@@ -26,7 +25,16 @@ class GenerosModelo {
     
         return $generos;
     }
- 
+
+    
+    public function obtenerLibros($id) {    
+        $query = $this->db->prepare('SELECT libros.titulo from generos inner join libros on generos.id = libros.id_genero  where   generos.id = ?');
+        $query->execute([$id]);   
+    
+        $libros = $query->fetchAll(PDO::FETCH_OBJ);
+        return $libros;
+    }
+    
     public function insertarGenero($nombre, $descripcion, $ruta_imagen ) { 
         $query = $this->db->prepare('INSERT INTO generos(nombre, descripcion, Ruta_imagen) VALUES (?, ?, ?)');
         $query->execute([$nombre, $descripcion, $ruta_imagen]);
@@ -37,11 +45,26 @@ class GenerosModelo {
     }
  
     public function borrarGenero($id) {
-        $query = $this->db->prepare('DELETE FROM generos WHERE id = ?');
-        $query->execute([$id]);
+        $libros = $this->obtenerLibros($id);
+        if (empty($libros)){
+
+            $genero = $this->obtenerGenero($id);
+            $query = $this->db->prepare('DELETE FROM generos WHERE id = ?');
+            $query->execute([$id]);   
+            if ($genero->Ruta_Imagen <> null){
+                unlink($genero->Ruta_Imagen);
+            }
+            return true;
+        }
+
+        return false;
+            
     }
 
-    public function actualizarGenero($id) {        
+    public function actualizarGenero($id , $nombre,$descripcion,$ruta_imagen) {    
+        echo "ENTREE";
+        $query = $this->db->prepare('UPDATE generos SET nombre= ? , descripcion=? , ruta_imagen = ? WHERE id = ?');
+        $query->execute([$nombre,$descripcion,$ruta_imagen,$id]);
         
     }
 }
